backend/requirements.txt
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
django-cors-headers==4.2.0
drf-spectacular==0.26.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
Pillow==10.0.0
factory-boy==3.3.0
django-cleanup==8.0.0
django-allauth==0.54.0
django-model-utils==4.3.1
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0
